2026.03.02
PostReport, PostLike에 Fk 삭제 및 분리
Redis - RedisTemplate, StringRedisTempalte 차이, Lettuce와 Jedis차이점, RDB와 AOF 장단점, 언제, 어디서 사용되고, 어떻게 사용하는지에 대해 공부중

+++ src/main/java/com/example/NoticeBoard/domain/post/service/PostLikeService.java
@@ -1,9 +1,12 @@
 package com.example.NoticeBoard.domain.post.service;
 
-import com.example.NoticeBoard.domain.post.entity.Post;
 import com.example.NoticeBoard.domain.post.entity.PostLike;
+import com.example.NoticeBoard.domain.post.repository.PostLikeRepository;
+import com.example.NoticeBoard.domain.post.repository.PostRepository;
+import com.example.NoticeBoard.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.data.redis.core.StringRedisTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -13,11 +16,17 @@ import org.springframework.transaction.annotation.Transactional;
 @Slf4j
 public class PostLikeService {
 
+    private final PostRepository postRepository;
+    private final PostLikeRepository postLikeRepository;
+    private final UserRepository userRepository;
+    private final StringRedisTemplate redisTemplate;
+
     // 게시글 좋아요
     public void likePost(Long postId, Long userId) {
 
-        Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new IllegalArgumentException("해당 게시글을 찾을 수 없습니다."));
+        if(!postRepository.existsById(postId)) {
+            throw new IllegalArgumentException("해당 게시글을 찾을 수 없습니다.");
+        }
 
         if(!userRepository.existsById(userId)){
             throw new IllegalArgumentException("해당 회원을 찾을 수 없습니다.");
@@ -30,20 +39,21 @@ public class PostLikeService {
 
         postLikeRepository.save(postLike);
 
-        post.setLikeCount(post.getLikeCount() + 1);
+        postRepository.incrementLikeCount(postId);
     }
 
     // 게시글 좋아요 취소
     public void unlikePost(Long postId, Long userId) {
 
-        Post post = postRepository.findById(postId)
-                .orElseThrow(() -> new IllegalArgumentException("게시글을 찾을 수 없습니다."));
+        if(!postRepository.existsById(postId)) {
+            throw new IllegalArgumentException("해당 게시글을 찾을 수 없습니다.");
+        }
 
         PostLike postLike = postLikeRepository.findByPostIdAndUserId(postId, userId)
                 .orElseThrow(() -> new IllegalStateException("좋아요 기록이 없습니다."));
 
         postLikeRepository.delete(postLike);
 
-        post.setLikeCount(post.getLikeCount() - 1);
+        postRepository.decrementLikeCount(postId);
     }
 }
